+++ draft-packages/guidance-block/KaizenDraft/GuidanceBlock/GuidanceBlock.tsx
@@ -1,10 +1,5 @@
-import {
-  Button,
-  ButtonProps,
-  Heading,
-  Icon,
-  Paragraph,
-} from "@kaizen/component-library"
+import { Button, ButtonProps } from "@kaizen/draft-button"
+import { Heading, Icon, Paragraph } from "@kaizen/component-library"
 import configureIcon from "@kaizen/component-library/icons/arrow-forward.icon.svg"
 import closeIcon from "@kaizen/component-library/icons/close.icon.svg"
 import classnames from "classnames"
@@ -28,6 +23,7 @@ type Props = {
     }
   }
   persistent?: boolean
+  variant?: "default" | "prominent"
   withActionButtonArrow?: boolean
 }
 
@@ -37,6 +33,7 @@ type State = {
 }
 class GuidanceBlock extends React.Component<Props, State> {
   static defaultProps = {
+    variant: "default",
     withActionButtonArrow: true,
   }
 
@@ -82,7 +79,9 @@ class GuidanceBlock extends React.Component<Props, State> {
     return (
       <div
         className={this.bannerClassName()}
-        style={{ marginTop: this.marginTop() }}
+        style={{
+          marginTop: this.marginTop(),
+        }}
         ref={this.containerRef}
         onTransitionEnd={this.onTransitionEnd}
       >
@@ -125,6 +124,7 @@ class GuidanceBlock extends React.Component<Props, State> {
   bannerClassName(): string {
     return classnames(styles.banner, {
       [styles.hidden]: this.state.hidden,
+      [styles.prominent]: this.props.variant === "prominent",
     })
   }