+++ draft-packages/events/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.71](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.70...@kaizen/draft-events@1.2.71) (2020-10-08)
+
+**Note:** Version bump only for package @kaizen/draft-events
+
+
+
+
+
+## [1.2.70](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.69...@kaizen/draft-events@1.2.70) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-events
+
+
+
+
+
+## [1.2.69](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.68...@kaizen/draft-events@1.2.69) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-events
+
+
+
+
+
 ## [1.2.68](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-events@1.2.67...@kaizen/draft-events@1.2.68) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-events