+++ draft-packages/filter-drawer/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-drawer@2.0.8...@kaizen/draft-filter-drawer@2.0.9) (2020-10-08)
+
+**Note:** Version bump only for package @kaizen/draft-filter-drawer
+
+
+
+
+
+## [2.0.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-drawer@2.0.7...@kaizen/draft-filter-drawer@2.0.8) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-filter-drawer
+
+
+
+
+
+## [2.0.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-drawer@2.0.6...@kaizen/draft-filter-drawer@2.0.7) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-filter-drawer
+
+
+
+
+
 ## [2.0.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-filter-drawer@2.0.5...@kaizen/draft-filter-drawer@2.0.6) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-filter-drawer