+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.4.1...@kaizen/draft-guidance-block@2.4.2) (2020-10-09)
+
+
+### Bug Fixes
+
+* Use correct props for button ([#811](https://github.com/cultureamp/kaizen-design-system/issues/811)) ([44de854](https://github.com/cultureamp/kaizen-design-system/commit/44de8545092364939d1ad89782ee160d975ccab5))
+
+
+
+
+
+## [2.4.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.4.0...@kaizen/draft-guidance-block@2.4.1) (2020-10-08)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+# [2.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.3.7...@kaizen/draft-guidance-block@2.4.0) (2020-10-07)
+
+
+### Features
+
+* GuidanceBlock Prominent Variant ([#804](https://github.com/cultureamp/kaizen-design-system/issues/804)) ([6818e35](https://github.com/cultureamp/kaizen-design-system/commit/6818e3513390df9235d1a35d3d737c83dca92326))
+
+
+
+
+
+## [2.3.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.3.6...@kaizen/draft-guidance-block@2.3.7) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [2.3.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.3.5...@kaizen/draft-guidance-block@2.3.6) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [2.3.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@2.3.4...@kaizen/draft-guidance-block@2.3.5) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block