+++ draft-packages/tooltip/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.44](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@1.5.43...@kaizen/draft-tooltip@1.5.44) (2020-10-08)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
+## [1.5.43](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@1.5.42...@kaizen/draft-tooltip@1.5.43) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
+## [1.5.42](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@1.5.41...@kaizen/draft-tooltip@1.5.42) (2020-10-01)
+
+**Note:** Version bump only for package @kaizen/draft-tooltip
+
+
+
+
+
 ## [1.5.41](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tooltip@1.5.40...@kaizen/draft-tooltip@1.5.41) (2020-09-30)
 
 **Note:** Version bump only for package @kaizen/draft-tooltip